% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_same.R
\name{order_same}
\alias{order_same}
\title{Auto change those pesky "Other please specify"s into "Other"}
\usage{
order_same(dataset, orders = ordered_df)
}
\arguments{
\item{dataset}{The name of the data frame that the mscharts pulls from, automatically included if function is piped in after running freqs. You almost never need any arguments in this function.}

\item{orders}{DEFAULT = ordered_df; First create an ordered dataframe called ordered_df using order_label. Your new dataframe crated by using order_same will have variables and groups in the same order as ordered_df}
}
\description{
Takes a dataframe (frequencies) and replaces the usual variations of "Other please specify" into Other. Converts all "None of the above" variations into "None of the above". Also removes all extra text in parantheses. Does this for both the 'label' and 'variable' vars.
}
\examples{
frequencies \%>\% order_same()
}
\keyword{extra}
\keyword{none}
\keyword{other}
