% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order Label Broken Down NSM.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Reading in lots of excel sheets in a workbook & turning it into one tibble}
\usage{
read_excel_allsheets(filename, tibble = TRUE)
}
\arguments{
\item{filename}{The file path to the .xlsx you want to pull all sheets from}

\item{tibble}{DEFAULT = TRUE; You will likely never change this}
}
\description{
Takes a dataframe (frequencies) and orders the labels and groups while adding percent labels for use in ggplot.
}
\examples{
responses <- read_excel_allsheets('~/filepath.xlsx')
}
