% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order Label Broken Down NSM.R
\name{topline}
\alias{topline}
\title{Add \%s to a topline report}
\usage{
topline(dataset,
  var_list = "place your variable names here with a \\"|\\" between them ")
}
\arguments{
\item{dataset}{The name of the data frame that the mscharts pulls from, usually piped in after running freqs. Please note that the variable column must be "variable" and the percentage column must be "result"}

\item{var_list}{If you only have variables that are percentages, you can leave this argument blank. Otherwise, add the names of you variables that are not percentages here separated by a "|". You do not have to type out the whole variable name. A unique portion of the var name will work as well.}
}
\description{
Takes a dataframe (frequencies) and for the first Y (result) of every X (variable), adds a \%. Also changes all 0 to <1 if n >=1
}
\examples{
frequencies \%>\% topline()
OR
frequencies \%>\% topline(
'DOLLARS_GIVEN|DONATIONS_RECEIVED')

}
\keyword{label}
\keyword{percent}
\keyword{topline}
