% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ages.R
\name{add_ages}
\alias{add_ages}
\title{Add age and census age groups to a data frame}
\usage{
add_ages(dataset, year_born_var, age_var)
}
\arguments{
\item{dataset}{The name of the data frame for the function to modify, usually piped in after running freqs. You almost never need any arguments in this function.}

\item{year_born_var}{DFAULT = NULL; You may either provide a year_born_var or an age_var, not both}

\item{age_var}{DFAULT = NULL}
}
\description{
Takes a dataframe with a year born variable or age variable and creates three new variables:
1. year_born_numeric
2. age_numeric
3. census_age_groups
}
\examples{
responses <- tibble::tibble(
  year_born = c(
  2000,
  2001
  ),
  age = c(
  20,
  19
  )
)

add_ages(responses, year_born_var = year_born)
add_ages(responses, age_var = age)
}
