% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_regions.R
\name{add_regions}
\alias{add_regions}
\title{Add a "census_region" variable based on states}
\usage{
add_regions(dataset, state_var)
}
\arguments{
\item{dataset}{The name of the data frame for the function to modify, usually piped in after running freqs. You almost never need any arguments in this function.}

\item{state_var}{NO DEFAULT; The state variable on which census regions will be based}
}
\description{
Takes a dataframe with a state variable and uses those states to categorize US census regions
}
\examples{
responses <- tibble::tibble(
  state = c(
  'Virginia',
  'UTAH',
  'New York',
  'alaska',
  'Washington DC'
  )
)

add_regions(responses, state_var = state)
}
