% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order Label Broken Down NSM.R
\name{order_label}
\alias{order_label}
\title{Order your data and add percent labels}
\usage{
order_label(dataset, label_var = label, inherent_order_label = F,
  group_var = F, inherent_order_group = F, label_specific = NA,
  group_specific = NA, stacked = F, ms_stacked = F, horizontal = F,
  rev_label = F, rev_group = F, l1 = label[1], g1 = group_var[1])
}
\arguments{
\item{dataset}{The name of the data frame that the mscharts pulls from, usually piped in after running freqs.}

\item{label_var}{DEFAULT = label; name of variable to be ordered}

\item{inherent_order_label}{DEFAULT = F; If F, puts labels in descending order. If T, puts labels in the inherent order from survey (Strongly agree to strongly disagree)}

\item{group_var}{DEFAULT = F; Add the name of the grouping variable if your data is grouped}

\item{inherent_order_group}{DEFAULT = F; If F, puts groups in descending order. If T, puts groups in the order they are factored (District 1, District 2...)}

\item{label_specific}{DEFAULT = NA; If T, puts the specified label first. ex: 'brand1' would put label called brand1 before all other labels}

\item{stacked}{DEFAULT = F; For stacked barcharts, use stacked = T (use ms_stacked for stacked mscharts)}

\item{ms_stacked}{DEFAULT = F; For stacked barcharts using mscharts, use ms_stacked = T. Specifying stacked = T automatically makes inherent_order_label = T}

\item{horizontal}{DEFAULT = F; For horizontal charts (grouped or ungrouped), use horizontal = T. Specifying ms_stacked = T automatically makes inherent_order_label = T}

\item{rev_label}{DEFAULT = F; To reverse the order of labels in a chart, use rev_label = T}

\item{rev_group}{DEFAULT = F; To reverse the order of groups in a chart, use rev_group = T}

\item{label_specific}{DEFAULT = NA; If T, puts the specified group first. ex: 'brand1' would put group called brand1 before all other groups}
}
\description{
Takes a dataframe (frequencies) and orders the labels and groups while adding percent labels for use in ggplot.
}
\examples{
frequencies \%>\% order_label(
horizontal = T,
inherent_order_label = T
)
OR
frequencies \%>\% order_label(
group_var = brand,
stacked = T,
group_specific = 'Y2 Analytics'
)
}
\keyword{dplyr::arrange}
\keyword{label}
\keyword{order}
