% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order Label Broken Down NSM.R
\name{other_rm}
\alias{other_rm}
\title{Auto change those pesky "Other please specify"s into "Other"}
\usage{
other_rm(dataset, var = label)
}
\arguments{
\item{dataset}{The name of the data frame that the mscharts pulls from, automatically included if function is piped in after running freqs.}

\item{var}{DEFAULT = label; Ideally, you never need any input in this function}
}
\description{
Takes a dataframe (frequencies) and replaces the usual variations of "Other please specify" into Other. Also converts all "None of the above" variations into only "None of the above".
}
\examples{
frequencies \%>\% other_rm()
}
\keyword{other}
